requirements.txt
flask==3.0.0
mysql-connector-python==9.5.0